Added a snapshot test examples and eslint rule change

+++ .eslintrc.js
@@ -3,7 +3,7 @@ module.exports = {
   plugins: ["react"],
   extends: "eslint:recommended",
   rules: {
-    quotes: [2, "double"],
+    quotes: [2, "single"],
     "react/jsx-uses-react": 1,
     "react/jsx-uses-vars": 1,
     "no-console": 2
